+++ src/php/array/array_key_last.ts
@@ -1,12 +1,4 @@
-import type { PhpArrayLike } from '../_helpers/_phpTypes.ts'
-
-const normalizeArrayKey = (key: string): string | number => {
-  if (/^(0|-?[1-9]\d*)$/.test(key)) {
-    return Number(key)
-  }
-
-  return key
-}
+import { normalizeArrayKey, type PhpArrayLike } from '../_helpers/_phpTypes.ts'
 
 export function array_key_last<T>(input: PhpArrayLike<T>): string | number | null {
   //      discuss at: https://locutus.io/php/array_key_last/